+++ tests/simulator/slli-01.cc
@@ -60,441 +60,441 @@ std::vector<uint8_t> slli_01::s_ScalarData;
 // Generate google tests
 
 TEST_F(slli_01, opcode__slli___op1_x17__dest_x27__op1val__0x40000001___immval_0x1d) {
-  ASSERT_EQ(getScalarWordData()[5], 3758096384);
+  ASSERT_EQ(getScalarWordData()[5], (uint32_t)3758096384);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x26__dest_x26__op1val_0x66666666___immval_0xf) {
-  ASSERT_EQ(getScalarWordData()[6], 858980352);
+  ASSERT_EQ(getScalarWordData()[6], (uint32_t)858980352);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x22__dest_x11__op1val__0x10001___immval_0x0) {
-  ASSERT_EQ(getScalarWordData()[7], 4294901759);
+  ASSERT_EQ(getScalarWordData()[7], (uint32_t)4294901759);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x15__dest_x6__op1val_0x4___immval_0x0) {
-  ASSERT_EQ(getScalarWordData()[8], 4);
+  ASSERT_EQ(getScalarWordData()[8], (uint32_t)4);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x9__dest_x16__op1val__0x400001___immval_0x1f) {
-  ASSERT_EQ(getScalarWordData()[9], 2147483648);
+  ASSERT_EQ(getScalarWordData()[9], (uint32_t)2147483648);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x11__dest_x20__op1val_0x4___immval_0x1f) {
-  ASSERT_EQ(getScalarWordData()[10], 0);
+  ASSERT_EQ(getScalarWordData()[10], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x1__dest_x19__op1val_0x8___immval_0x8) {
-  ASSERT_EQ(getScalarWordData()[11], 2048);
+  ASSERT_EQ(getScalarWordData()[11], (uint32_t)2048);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x19__dest_x25__op1val__0x80000000___immval_0x10) {
-  ASSERT_EQ(getScalarWordData()[12], 0);
+  ASSERT_EQ(getScalarWordData()[12], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x8__dest_x12__op1val_0x0___immval_0xc) {
-  ASSERT_EQ(getScalarWordData()[13], 0);
+  ASSERT_EQ(getScalarWordData()[13], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x27__dest_x30__op1val_0x7fffffff___immval_0x8) {
-  ASSERT_EQ(getScalarWordData()[14], 4294967040);
+  ASSERT_EQ(getScalarWordData()[14], (uint32_t)4294967040);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x2__dest_x4__op1val_0x1___immval_0x1) {
-  ASSERT_EQ(getScalarWordData()[15], 2);
+  ASSERT_EQ(getScalarWordData()[15], (uint32_t)2);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x31__dest_x14__op1val_0x2___immval_0x6) {
-  ASSERT_EQ(getScalarWordData()[16], 128);
+  ASSERT_EQ(getScalarWordData()[16], (uint32_t)128);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x24__dest_x17__op1val_0x10___immval_0xe) {
-  ASSERT_EQ(getScalarWordData()[17], 262144);
+  ASSERT_EQ(getScalarWordData()[17], (uint32_t)262144);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x4__dest_x10__op1val_0x20___immval_0x3) {
-  ASSERT_EQ(getScalarWordData()[18], 256);
+  ASSERT_EQ(getScalarWordData()[18], (uint32_t)256);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x18__dest_x2__op1val_0x40___immval_0x15) {
-  ASSERT_EQ(getScalarWordData()[19], 134217728);
+  ASSERT_EQ(getScalarWordData()[19], (uint32_t)134217728);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x5__dest_x23__op1val_0x80___immval_0x15) {
-  ASSERT_EQ(getScalarWordData()[20], 268435456);
+  ASSERT_EQ(getScalarWordData()[20], (uint32_t)268435456);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x13__dest_x8__op1val_0x100___immval_0x1) {
-  ASSERT_EQ(getScalarWordData()[21], 512);
+  ASSERT_EQ(getScalarWordData()[21], (uint32_t)512);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x20__dest_x0__op1val_0x200___immval_0x0) {
-  ASSERT_EQ(getScalarWordData()[22], 0);
+  ASSERT_EQ(getScalarWordData()[22], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x16__dest_x9__op1val_0x400___immval_0x2) {
-  ASSERT_EQ(getScalarWordData()[23], 4096);
+  ASSERT_EQ(getScalarWordData()[23], (uint32_t)4096);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x21__dest_x5__op1val_0x800___immval_0x13) {
-  ASSERT_EQ(getScalarWordData()[24], 1073741824);
+  ASSERT_EQ(getScalarWordData()[24], (uint32_t)1073741824);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x23__dest_x1__op1val_0x1000___immval_0x7) {
-  ASSERT_EQ(getScalarWordData()[25], 524288);
+  ASSERT_EQ(getScalarWordData()[25], (uint32_t)524288);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x12__dest_x18__op1val_0x2000___immval_0x10) {
-  ASSERT_EQ(getScalarWordData()[26], 536870912);
+  ASSERT_EQ(getScalarWordData()[26], (uint32_t)536870912);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x29__dest_x15__op1val_0x4000___immval_0xb) {
-  ASSERT_EQ(getScalarWordData()[27], 33554432);
+  ASSERT_EQ(getScalarWordData()[27], (uint32_t)33554432);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x3__dest_x21__op1val_0x8000___immval_0x17) {
-  ASSERT_EQ(getScalarWordData()[28], 0);
+  ASSERT_EQ(getScalarWordData()[28], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x0__dest_x31__op1val_0x0___immval_0x1) {
-  ASSERT_EQ(getScalarWordData()[29], 0);
+  ASSERT_EQ(getScalarWordData()[29], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x14__dest_x3__op1val_0x20000___immval_0x1b) {
-  ASSERT_EQ(getScalarWordData()[30], 0);
+  ASSERT_EQ(getScalarWordData()[30], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x25__dest_x24__op1val_0x40000___immval_0x1f) {
-  ASSERT_EQ(getScalarWordData()[31], 0);
+  ASSERT_EQ(getScalarWordData()[31], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x30__dest_x29__op1val_0x80000___immval_0xf) {
-  ASSERT_EQ(getScalarWordData()[32], 0);
+  ASSERT_EQ(getScalarWordData()[32], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x28__dest_x13__op1val_0x100000___immval_0x17) {
-  ASSERT_EQ(getScalarWordData()[33], 0);
+  ASSERT_EQ(getScalarWordData()[33], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x7__op1val_0x200000___immval_0x7) {
-  ASSERT_EQ(getScalarWordData()[34], 268435456);
+  ASSERT_EQ(getScalarWordData()[34], (uint32_t)268435456);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x7__dest_x22__op1val_0x400000___immval_0x1d) {
-  ASSERT_EQ(getScalarWordData()[35], 0);
+  ASSERT_EQ(getScalarWordData()[35], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x6__dest_x28__op1val_0x800000___immval_0x2) {
-  ASSERT_EQ(getScalarWordData()[36], 33554432);
+  ASSERT_EQ(getScalarWordData()[36], (uint32_t)33554432);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x1000000___immval_0x6) {
-  ASSERT_EQ(getScalarWordData()[37], 1073741824);
+  ASSERT_EQ(getScalarWordData()[37], (uint32_t)1073741824);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x2000000___immval_0x5) {
-  ASSERT_EQ(getScalarWordData()[38], 1073741824);
+  ASSERT_EQ(getScalarWordData()[38], (uint32_t)1073741824);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x4000000___immval_0x11) {
-  ASSERT_EQ(getScalarWordData()[39], 0);
+  ASSERT_EQ(getScalarWordData()[39], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x8000000___immval_0x0) {
-  ASSERT_EQ(getScalarWordData()[40], 134217728);
+  ASSERT_EQ(getScalarWordData()[40], (uint32_t)134217728);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x10000000___immval_0x3) {
-  ASSERT_EQ(getScalarWordData()[41], 2147483648);
+  ASSERT_EQ(getScalarWordData()[41], (uint32_t)2147483648);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x20000000___immval_0xc) {
-  ASSERT_EQ(getScalarWordData()[42], 0);
+  ASSERT_EQ(getScalarWordData()[42], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x40000000___immval_0xb) {
-  ASSERT_EQ(getScalarWordData()[43], 0);
+  ASSERT_EQ(getScalarWordData()[43], (uint32_t)0);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x2___immval_0xe) {
-  ASSERT_EQ(getScalarWordData()[44], 4294934528);
+  ASSERT_EQ(getScalarWordData()[44], (uint32_t)4294934528);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x3___immval_0x1) {
-  ASSERT_EQ(getScalarWordData()[45], 4294967290);
+  ASSERT_EQ(getScalarWordData()[45], (uint32_t)4294967290);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x5___immval_0x7) {
-  ASSERT_EQ(getScalarWordData()[46], 4294966656);
+  ASSERT_EQ(getScalarWordData()[46], (uint32_t)4294966656);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x9___immval_0xa) {
-  ASSERT_EQ(getScalarWordData()[47], 4294958080);
+  ASSERT_EQ(getScalarWordData()[47], (uint32_t)4294958080);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x11___immval_0x11) {
-  ASSERT_EQ(getScalarWordData()[48], 4292739072);
+  ASSERT_EQ(getScalarWordData()[48], (uint32_t)4292739072);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x21___immval_0xe) {
-  ASSERT_EQ(getScalarWordData()[49], 4294426624);
+  ASSERT_EQ(getScalarWordData()[49], (uint32_t)4294426624);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x41___immval_0x13) {
-  ASSERT_EQ(getScalarWordData()[50], 4260888576);
+  ASSERT_EQ(getScalarWordData()[50], (uint32_t)4260888576);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x81___immval_0x10) {
-  ASSERT_EQ(getScalarWordData()[51], 4286513152);
+  ASSERT_EQ(getScalarWordData()[51], (uint32_t)4286513152);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x101___immval_0xb) {
-  ASSERT_EQ(getScalarWordData()[52], 4294440960);
+  ASSERT_EQ(getScalarWordData()[52], (uint32_t)4294440960);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x201___immval_0x11) {
-  ASSERT_EQ(getScalarWordData()[53], 4227727360);
+  ASSERT_EQ(getScalarWordData()[53], (uint32_t)4227727360);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x401___immval_0x7) {
-  ASSERT_EQ(getScalarWordData()[54], 4294836096);
+  ASSERT_EQ(getScalarWordData()[54], (uint32_t)4294836096);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x801___immval_0xf) {
-  ASSERT_EQ(getScalarWordData()[55], 4227825664);
+  ASSERT_EQ(getScalarWordData()[55], (uint32_t)4227825664);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x1001___immval_0xa) {
-  ASSERT_EQ(getScalarWordData()[56], 4290771968);
+  ASSERT_EQ(getScalarWordData()[56], (uint32_t)4290771968);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x2001___immval_0x6) {
-  ASSERT_EQ(getScalarWordData()[57], 4294442944);
+  ASSERT_EQ(getScalarWordData()[57], (uint32_t)4294442944);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x4001___immval_0x2) {
-  ASSERT_EQ(getScalarWordData()[58], 4294901756);
+  ASSERT_EQ(getScalarWordData()[58], (uint32_t)4294901756);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x8001___immval_0x3) {
-  ASSERT_EQ(getScalarWordData()[59], 4294705144);
+  ASSERT_EQ(getScalarWordData()[59], (uint32_t)4294705144);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x20001___immval_0x1d) {
-  ASSERT_EQ(getScalarWordData()[60], 3758096384);
+  ASSERT_EQ(getScalarWordData()[60], (uint32_t)3758096384);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x40001___immval_0x1d) {
-  ASSERT_EQ(getScalarWordData()[61], 3758096384);
+  ASSERT_EQ(getScalarWordData()[61], (uint32_t)3758096384);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x80001___immval_0xc) {
-  ASSERT_EQ(getScalarWordData()[62], 2147479552);
+  ASSERT_EQ(getScalarWordData()[62], (uint32_t)2147479552);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x200001___immval_0x6) {
-  ASSERT_EQ(getScalarWordData()[63], 4160749504);
+  ASSERT_EQ(getScalarWordData()[63], (uint32_t)4160749504);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x800001___immval_0x13) {
-  ASSERT_EQ(getScalarWordData()[64], 4294443008);
+  ASSERT_EQ(getScalarWordData()[64], (uint32_t)4294443008);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x1000001___immval_0x11) {
-  ASSERT_EQ(getScalarWordData()[65], 4294836224);
+  ASSERT_EQ(getScalarWordData()[65], (uint32_t)4294836224);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x2000001___immval_0xf) {
-  ASSERT_EQ(getScalarWordData()[66], 4294934528);
+  ASSERT_EQ(getScalarWordData()[66], (uint32_t)4294934528);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x4000001___immval_0xd) {
-  ASSERT_EQ(getScalarWordData()[67], 4294959104);
+  ASSERT_EQ(getScalarWordData()[67], (uint32_t)4294959104);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x8000001___immval_0xe) {
-  ASSERT_EQ(getScalarWordData()[68], 4294950912);
+  ASSERT_EQ(getScalarWordData()[68], (uint32_t)4294950912);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x10000001___immval_0x7) {
-  ASSERT_EQ(getScalarWordData()[69], 4294967168);
+  ASSERT_EQ(getScalarWordData()[69], (uint32_t)4294967168);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x20000001___immval_0xb) {
-  ASSERT_EQ(getScalarWordData()[70], 4294965248);
+  ASSERT_EQ(getScalarWordData()[70], (uint32_t)4294965248);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x55555555___immval_0x11) {
-  ASSERT_EQ(getScalarWordData()[71], 2863267840);
+  ASSERT_EQ(getScalarWordData()[71], (uint32_t)2863267840);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x55555556___immval_0x17) {
-  ASSERT_EQ(getScalarWordData()[72], 1426063360);
+  ASSERT_EQ(getScalarWordData()[72], (uint32_t)1426063360);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x3___immval_0x13) {
-  ASSERT_EQ(getScalarWordData()[73], 1572864);
+  ASSERT_EQ(getScalarWordData()[73], (uint32_t)1572864);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x5___immval_0x1) {
-  ASSERT_EQ(getScalarWordData()[74], 10);
+  ASSERT_EQ(getScalarWordData()[74], (uint32_t)10);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x33333333___immval_0x3) {
-  ASSERT_EQ(getScalarWordData()[75], 2576980376);
+  ASSERT_EQ(getScalarWordData()[75], (uint32_t)2576980376);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0xb504___immval_0x6) {
-  ASSERT_EQ(getScalarWordData()[76], 4292001536);
+  ASSERT_EQ(getScalarWordData()[76], (uint32_t)4292001536);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x66666667___immval_0x17) {
-  ASSERT_EQ(getScalarWordData()[77], 864026624);
+  ASSERT_EQ(getScalarWordData()[77], (uint32_t)864026624);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0xb503___immval_0x6) {
-  ASSERT_EQ(getScalarWordData()[78], 4292001600);
+  ASSERT_EQ(getScalarWordData()[78], (uint32_t)4292001600);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0xb505___immval_0x1f) {
-  ASSERT_EQ(getScalarWordData()[79], 2147483648);
+  ASSERT_EQ(getScalarWordData()[79], (uint32_t)2147483648);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x33333332___immval_0x4) {
-  ASSERT_EQ(getScalarWordData()[80], 858993440);
+  ASSERT_EQ(getScalarWordData()[80], (uint32_t)858993440);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x6___immval_0x1e) {
-  ASSERT_EQ(getScalarWordData()[81], 2147483648);
+  ASSERT_EQ(getScalarWordData()[81], (uint32_t)2147483648);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0xb504___immval_0x7) {
-  ASSERT_EQ(getScalarWordData()[82], 5931520);
+  ASSERT_EQ(getScalarWordData()[82], (uint32_t)5931520);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x100001___immval_0x8) {
-  ASSERT_EQ(getScalarWordData()[83], 4026531584);
+  ASSERT_EQ(getScalarWordData()[83], (uint32_t)4026531584);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x55555554___immval_0x3) {
-  ASSERT_EQ(getScalarWordData()[84], 2863311520);
+  ASSERT_EQ(getScalarWordData()[84], (uint32_t)2863311520);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x66666665___immval_0x10) {
-  ASSERT_EQ(getScalarWordData()[85], 1717895168);
+  ASSERT_EQ(getScalarWordData()[85], (uint32_t)1717895168);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0xb503___immval_0x10) {
-  ASSERT_EQ(getScalarWordData()[86], 3036872704);
+  ASSERT_EQ(getScalarWordData()[86], (uint32_t)3036872704);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x55555556___immval_0x17) {
-  ASSERT_EQ(getScalarWordData()[87], 2868903936);
+  ASSERT_EQ(getScalarWordData()[87], (uint32_t)2868903936);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val__0x55555555___immval_0xe) {
-  ASSERT_EQ(getScalarWordData()[88], 2863316992);
+  ASSERT_EQ(getScalarWordData()[88], (uint32_t)2863316992);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x6___immval_0x1e) {
-  ASSERT_EQ(getScalarWordData()[89], 2147483648);
+  ASSERT_EQ(getScalarWordData()[89], (uint32_t)2147483648);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x33333334___immval_0x1d) {
-  ASSERT_EQ(getScalarWordData()[90], 2147483648);
+  ASSERT_EQ(getScalarWordData()[90], (uint32_t)2147483648);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x200___immval_0x0) {
-  ASSERT_EQ(getScalarWordData()[91], 512);
+  ASSERT_EQ(getScalarWordData()[91], (uint32_t)512);
 }
 
 
 TEST_F(slli_01, opcode__slli___op1_x10__dest_x11__op1val_0x10000___immval_0x1) {
-  ASSERT_EQ(getScalarWordData()[92], 131072);
+  ASSERT_EQ(getScalarWordData()[92], (uint32_t)131072);
 }
